Moved App.tsx to pages folder
